move mirage application serializer to addon to avoid conflicts
