+++ mirage-support/serializers/github-pull.js
@@ -1,4 +1,4 @@
-import ApplicationSerializer from './application';
+import ApplicationSerializer from './addon';
 
 export default ApplicationSerializer.extend({
   include: ['user']
