+++ mirage-support/serializers/github-release.js
@@ -1,4 +1,4 @@
-import ApplicationSerializer from './application';
+import ApplicationSerializer from './addon';
 
 export default ApplicationSerializer.extend({
   include: ['author']
